Game: Document tile map transformation
Rather cleverly, this game uses cells in a tile map to represent the starting
position of the player and any enemies. When the level is loaded, the game
instantiates the player and enemies at the positions specified by the cells in
the tile map, then removes them from the tile map.

Document this a little more clearly for beginners. Also explain why the wall
cells are modified: this is purely visual.

Fixes https://github.com/endlessm/everlasting-candy/issues/4