Count levels dynamically
Previously, the number of levels was hardcoded in the Global script, and would
have to match the number of scene files in the TileMap directory.

Instead, when Global is initialized, enumerate the numbered scene files in that
directory; sort them numerically; and load them for good measure. The number of
levels is then determined from the number of scenes found.

Fixes https://github.com/endlessm/everlasting-candy/issues/5