Add world selector
Allow learners to contribute their own sequences of levels by introducing the
concept of worlds.

Repurpose the original game as the “Candy Wrapper” world, renaming its directory
accordingly.

Add a new main scene
which provides the option to either play the original game, or one of a series
of additional worlds. Enumerate worlds in a particular directory of the source
tree – a world is just defined to be a subdirectory of `res://Worlds`, whose
folder name becomes the name of the world, and which is expected (like the
original game) to contain a sequence of numbered `TileMapLayer` scenes.

Add a Sample world within that folder which contains the first and last
pseudo-levels of the original game, plus two levels with crude recreation of
logos close to our hearts and one level with an actual heart (designed by my
in-house playtester). Renumber the closing pseudo-level to 999 to make it easier
to add worlds in between.

The original game's titles are hand-lettered in Aseprite. On the new world
selector screen, use a pixel font I bought in a bundle on itch.io some time ago.
Following its CC 4.0 Attribution license, add attribution to the README.

Fixes https://github.com/endlessm/everlasting-candy/issues/11