Game: Return to menu when Esc is pressed
This allows the player to choose a different world. Because it is now possible
to leave and re-enter the game proper, we must reset the global.level variable
when entering a world.

It may be nice to save the player's progress in a world, but this is a short,
simple game so it does not seem worth the complexity.

In addition, only play music while in-game.