Fix loading levels in exported game
When the game is exported, textual resource files (e.g.
res://CandyWrapper/0.tscn) are converted to a binary representation. When you
use the DirAccess API to list the files in a directory, you see the names of the
binary files (e.g. res://CandyWrapper/0.tscn.remap). These do not match the
regular expression in our code, and so no levels are loaded.

In addition, in order to actually load the level scenes, the original path (e.g.
res://CandyWrapper.0.tscn) must be used.

Godot 4.4 will (unless it is removed before release) add a
ResourceLoader.list_directory() function which does what we need here: to
enumerate a directory and feast on the resources within.

For now, add a workaround that accepts and strips any .remap suffix found while
listing the directory. Remove the regular expression that required scenes to be
numerically-named: I don't think it really adds anything. We could reintroduce
it if I'm proven wrong.

Fixes https://github.com/endlessm/everlasting-candy/issues/24