ExtraWorldSelector: Add scrolling
Once we have 4 extra worlds, they will not fit vertically into the game window.

Add vertical scrolling. Include the whole page in the scroll, including the back
button and title, so that as many worlds as possible fit. Ensure the focused
button is always in view, so that they can be navigated with a gamepad or arrow
keys.

The scrollbar is set to be 4 pixels wide, and always visible; the right margin
of the page is adjusted to make room for it. This is ugly while we don't need
it, but simpler than trying to hide it when it's not needed.

Fixes https://github.com/endlessm/everlasting-candy/issues/26