Player: Emit signals on contact with goobers
Previously, Player would directly call methods on the Game scene instance
(accessed via a global variable) when it either stomps on an enemy, or dies.

Instead, emit signals from Player and connect to them from Game to run the same
logic.

Since this was the only use of the global variable, remove it.