Introduce World scene for each world
Previously, the top-level scene would either be WorldSelector or Game.
To transition between levels, Game would adjust a global variable
containing the current level number and then reload itself. When Game
loads, it accesses that global variable and instantiates the
corresponding map.

This commit introduces a new World scene for each world, using a shared
script.

When World is loaded, it loads all levels in its own directory, and then
instantiates the Game scene as a child, setting the first level as its
map. When the player wins or loses a level, Game now emits a signal. In
response, World replaces its child Game scene with a new instance, with
the appropriate map. Game grows a new property to trigger the special
behaviour for the first and last level, rather than checking global
variables for this. As a result, Game no longer accesses global
variables.

CandySpawner moves from the Game scene to the World, so that World can
adjust the spawn rate based on progress through the worlds. As a result,
CandySpawner no longer accesses global variables. This also opens the
possibility of different worlds having a different texture falling in
the background – or none at all.

The background music moves from Global to World. Similarly, this allows
different worlds to have different background music.

Game has a simple Map so that the Game scene can be run in the editor.
In due course, I plan to eliminate the Game scene, and instead make
every level a playable scene in its own right.