Parametrize background candy
Define a CandySpawner scene, which is nothing more than a Node2D with a
script attached. Replace the Node2D with an instance of that scene in
the two Worlds. Add an exported property to the new scene for the candy
texture, defaulting to the current texture.

This allows a world to have something different falling in the
background.