Move win/lose sounds & text overlay to World
Previously, the Game scene contained the AudioStreamPlayer nodes for the
"win" and "lose" sounds, and a Sprite2D that contains the text overlay
for the "level complete" and "lose" messages, plus the title screen and
game-complete screen. The Game would only signal that the level is
complete or that the player has died after a delay.

Move these nodes out to the World, making them customisable for
different worlds. Move the delay into World, using a transient Timer
node to simplify the logic. Move the logic that advances to the next
level immediately when the player presses the jump key on the title or
complete screens out to the world, too.