Make each level a playable scene
Previously, each level was a scene consisting of a single TileMapLayer.
At runtime, this was loaded and added as a child of the Game scene.
Among other things, this made it hard to test a level in isolation: it
couldn't be run by itself.

Rename the Game script to Level. Rewrite every level to have the same
structure that the Game scene had:

    Level: Node2D (with Level.gd attached)
      - Map: TileMapLayer
      - Goobers: Node2D

Set the level_type property on the first and last levels of each world.

Rework World.gd to remove the magical manipulation of the Game scene,
and instead add the level directly as a child.

Remove the now-unused Game.tscn.