Improve README and split out world building doc
Split out the world building doc into its own file - it is going
to be expanded with more content, and I don't want to flood the README.

Adjust the README to shift the focus in the way we present the project.
The fact that it is used in our learning programs is true and interesting,
but the focus should be on this being a real game.

The world building documentation was adjusted to have long/unwrapped
lines, to be consistent with the other docs. I don't feel strongly but
https://code.dblock.org/2021/06/07/to-wrap-or-not-to-wrap-in-markdown.html
is a pretty good argument for this.

Add a link to the GDD too.