Level: Parametrise the Explosion scene
The Level script instantiates an explosion when the player and a goober
come into contact. The scene to instantiate is stored in an instance variable.

Mark this variable with @export to make it easier to override in the
editor, without having to modify or extend the Level script. Switch from
load() to preload() since the default path is a constant.

With this change, it is easier to, for example, create a new world where
the “explosion” is a shower of magical sparkles rather than an explosion
in the violent sense.