Level: Count goobers by group
Previously, the Level would check whether any goobers remained by
counting the children of a Goobers node.

Instead, count members of a new `goober` group. This reduces the
requirement for a level to have a specific node hierarchy to work
correctly: if the level designer instantiates a goober directly into the
scene, somewhere other than as a child of a node called Goobers, the
level will still work correctly.