Level: Bind to Player signals by group
Previously, when the Level script instantiated a Player scene based on a
tile in the TileMapLayer, its signals would be connected to handlers in
the Level class; but it was not easy to connect the signals on a Player
scene that had been instantiated into the level scene in the editor,
because the Player.died signal doesn't include a reference to the
emitting player, which the Level callback needs.

Create a new global group, `player`, and add the Player scene to that
group. Then, in `Level._ready()`, after the TileMapLayer has been
processed, connect to those signals on any members of that group which
are in the tree.

(This preserves the previous behaviour where, if you place two or more
players into the tilemap, they are all "live". I can imagine some
interesting level designs based on this behaviour!)