Add player & goober scenes to TileSet
Source IDs 1 and 2 in the TileSet used by each level's TileMapLayer are
“magic”: there is code in Level.gd which finds cells with these source
IDs, and replaces them with instantiated scenes. The scenes to
instantiate are hardcoded in Level.gd.

While it's most common to use a TileSetAtlasSource to create tiles from
a texture, it is possible to use a TileSetSceneCollectionSource where
tiles are instances of scenes. At runtime, they are instantiated as
children of the TileMapLayer. In effect, this is the
"batteries-included" version of what Level.gd is doing manually.

Define a new source in the shared tileset that includes the Player and
Goober scenes.

Adjust the code in Level.gd to connect signals to any players that are
added as children of the TileMapLayer after the scene becomes _ready.
This is because (from the TileMapLayer documentation):

> For performance reasons, all TileMap updates are batched at the end of
> a frame. Notably, this means that scene tiles from a
> TileSetScenesCollectionSource may be initialized after their parent.
> This is only queued when inside the scene tree.

The documentation suggests that one can use
TileMapLayer.update_internals() to force an earlier update, but, in my
testing, calling this from `Level._ready()` causes the following
sequence of events:

- Level._ready()
  - TileMapLayer.update_internals()
    - Player added as child of TileMapLayer
- At end of frame:
  - Another Player is instantiated as child of TileMapLayer
  - Original Player is freed

So this only makes matters worse.

This commit does not update any levels' maps to use this new TileSet
source; that will come later.

https://docs.godotengine.org/en/stable/classes/class_tilesetscenescollectionsource.html
https://docs.godotengine.org/en/stable/classes/class_tilemaplayer.html