Use Object.connect() for player signals
Since commit bf38e462 the player signals are connected dynamically by
traversing the group. However this fails at runtime with:

"Invalid access to property or key 'died' on a base object of type
'CharacterBody2D'."

This is because signal.connect() is used. That is the preferred way,
except it doesn't work dynamically. So use Object.connect() passing the
signal StringName instead.