Revert "Use Object.connect() for player signals"
This reverts commit 90c7259470307712e9c3ea995db25bbe39c647eb.

The issue was that the Player scene instantiated in title/you-win levels
had its script removed so that it did not respond to input and just
played its idle animation. As a result, it did not have those signals
defined.

Commit 90c7259470307712e9c3ea995db25bbe39c647eb turned a crash at
runtime into errors printed to the error log:

    E 0:00:11:0309   Level.gd:34 @ _ready(): In Object of type 'CharacterBody2D': Attempt to connect nonexistent signal 'died' to callable 'Node2D(Level.gd)::_on_died'.
      <C++ Error>    Condition "!signal_is_valid" is true. Returning: ERR_INVALID_PARAMETER
      <C++ Source>   core/object/object.cpp:1390 @ connect()
      <Stack Trace>  Level.gd:34 @ _ready()
                     World.gd:67 @ _instantiate_level()
                     World.gd:57 @ _ready()

Instead, fully-functional Player instances are now placed into these
scenes, with the new inert property set to disable interactivity. This
means they have the normal signals, even if they will never be emitted.