World scenes: Add easy to mod background
Add a ColorRect node named "Background" to each existing World scene, at
the bottom of the hierarchy. The color is set to the same as in the
default environment clear color in Project Settings.

This allows easy tweak of the background color in a world, which will be
documented in next commit.

https://github.com/endlessm/everlasting-candy/issues/63