Level: Add goobers directly to scene
Prior to commit 3e259e2a23699ad46d831cbec3ab03598bf7149a ("Level: Count
goobers by group"), each Goober scene would have to be a child of the
Goobers node to be counted.

Now, since they are counted by membership of the `goobers` group, they
can appear anywhere in the scene.

Add Goobers directly to the level. Remove the variable that points to
the Goobers node. Remove the Goobers nodes from every level in the tree.
Remove the reference to it from the world-building documentation.